Newsletter/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
feedparser==6.0.10
openai==1.3.0
python-dotenv==1.0.0
schedule==1.2.0
jinja2==3.1.2
markdown==3.5.1
pandas==2.1.3
nltk==3.8.1
textblob==0.17.1
newspaper3k==0.2.8
selenium==4.15.2
webdriver-manager==4.0.1
tweepy==4.14.0
praw==7.7.1 